Fix and improve the debug build

+++ src/ldcache.h
@@ -19,7 +19,7 @@ struct ldcache {
 };
 
 enum {
-        LD_UNKNOWN                 = (uint32_t)-1,
+        LD_UNKNOWN                 = -1,
 
         LD_TYPE_MASK               = 0x00ff,
         LD_ELF                     = 0x0001,
